utils for semantics
